==> MultiClass-Classification/Final-Annotation/ECCO 1/Step0-DataAnalysis.ipynb <==
## Some data analysis before the annotation process:

import os
import pandas as pd
import numpy as np

# Checking if we have same category names:
path = '../txtLAB/Doc Class 5.0/ECCO_TrainingData_Trial4/'

print(len(os.listdir(path)[2:]))

categories = ['Agriculture.csv', 'Botany.csv', 'Church.csv', 'Commerce.csv', 
              'Drama.csv', 'Fiction.csv', 'History.csv', 'History_Natural.csv', 'Law.csv', 'Mathematics.csv', 
              'Medicine.csv', 'Physics.csv', 'Poetry.csv', 'Politics.csv', 'Rhetoric.csv', 'Sermons.csv', 
              'Travels.csv']
print(set(categories) == set(os.listdir(path)[2:]))

#### Creating a table for all 17 classes:

df_s = []
for cat in os.listdir(path)[2:]:
    df = pd.read_csv(path+cat)
    df['Class Label'] = cat[:-4]
    print(cat)
    print(df.shape)
    df_s.append(df)

table1_all_17 = pd.concat(df_s, ignore_index=True)
print(table1_all_17.nunique())
table1_all_17.head()

table1_all_17['Class Label'].value_counts()

table1_all_17.to_csv('/Users/dh_lab_05/Desktop/FINAL_Labelling/training_data.csv', index=None)

check = pd.read_csv('./training_data.csv')
check.head()

check.nunique()

#### Creating a table for ALL the data we have:

path = '../txtLAB/Koustuv_ECCO_Metadata/'

df_s2 = []
for cat in os.listdir(path):
    df = pd.read_csv(path+cat, delimiter='\t')
    df['TableName'] = cat[:-4]
    print(cat)
    print(df.shape)
    df_s2.append(df)

table2_all_metadata = pd.concat(df_s2, ignore_index=True)
print(table2_all_metadata.nunique())
table2_all_metadata.head()

#### This means, I actually have to label 97688 (154924 - 57236) files.
#### Creating a CSV that has only the metadata I need to actually label:

filenames_already_labelled = table1_all_17['Filename'].tolist()
print(len(filenames_already_labelled))

to_label = table2_all_metadata.loc[~table2_all_metadata['Filename'].isin(filenames_already_labelled)]

to_label.nunique()

to_label.to_csv('./fnames_to_label.csv', index=None)

temp = pd.read_csv('./fnames_to_label.csv')
temp.head()

temp.nunique()

# Sanity check: Checking if metadata had those 57k files used for training SVM:
print(len(filenames_already_labelled))
blah = table2_all_metadata.loc[table2_all_metadata['Filename'].isin(filenames_already_labelled)]
print(blah.nunique())
blah.head()
