==> MultiClass-Classification/Final-Annotation/ECCO 1/Step1-Pickle-testFeatures-ECCO1.ipynb <==
## Calculating feature vectors for the files that I've to label:

import pandas as pd
to_label = pd.read_csv('./fnames_to_label.csv')
print(to_label.nunique())
to_label.head()

df = to_label[['Filename', 'TableName']]
df.head()

set(df['TableName'].tolist())

# Model 2: All words > len(3)

from nltk.tokenize import word_tokenize, sent_tokenize
from collections import defaultdict
from heapq import nlargest

'''
Takes in a list of sentences where each sentence is a list of words, and optional argument 'user_stopwords'.
Returns a dictionary with each 'word' is the key, and 'count' as the value.
'''
def calculate_frequencies(sentences_ll, user_stopwords=None):  # sentences_ll is a list of lists
    frequency = defaultdict(int)    # default value : 0
    
    for sentence in sentences_ll:
        for word in sentence:
            word = word.lower()
            
            if len(word) > 3:
                frequency[word] += 1

    return frequency

'''
Takes in text, and n = number of features
Returns a list of n most frequent words
'''
def get_features(text, n, user_stopwords=None):  # n is the desired no. of features
    sentences = sent_tokenize(text.decode('utf8'))
    
    sentences_ll = []
    for s in sentences:
        words = word_tokenize(s)
        sentences_ll.append(words)

    frequency = calculate_frequencies(sentences_ll, user_stopwords)
    return nlargest(n, frequency, key=frequency.get)

tuples = [tuple(x) for x in df.values]
print(tuples[:4])

# Read files from Dataset
path = './FINAL_Dataset/'

map_filename_to_1000words = {}

counter = 0
for (fname, folder) in tuples[:100]:
    with open(path+folder+'/'+fname+'.txt', 'rb') as f:
        map_filename_to_1000words[fname] = get_features(f.read(), 1000)
        
    if counter % 10 == 0:
        print(counter)
    counter += 1
    
# Should stop at 95000
# Started running at 3PM Tuesday

import pickle

with open('./features_97k.pickle', 'wb') as f:
    pickle.dump(map_filename_to_1000words, f)

with open('./features_97k.pickle', 'rb') as f:
    df = pickle.load(f)
    
print(df.keys())
