==> MultiClass-Classification/Final-Annotation/ECCO 1/Step3-CreateFinalCSV.ipynb <==
## Using the output pickle of Step2, and adding the ClassLabel column to metadata:

import pickle
path = '../Map_FileToLabel.pickle'

with open(path, 'rb') as f:
    FileToLabel = pickle.load(f)

print(len(FileToLabel))

print(type(FileToLabel))

# print(FileToLabel)

# Match it with ECCO metadata:
import pandas as pd
import os
import numpy as np

path = '../txtLAB/Koustuv_ECCO_Metadata/'

df_s = []
for cat in os.listdir(path):
    df = pd.read_csv(path+cat, delimiter='\t')
    df['TableName'] = cat[:-4]
    print(cat)
    print(df.shape)
    df_s.append(df)

table2_all_metadata = pd.concat(df_s, ignore_index=True)
print(table2_all_metadata.nunique())
table2_all_metadata.head()

fnames_to_label = list(FileToLabel.keys())
print(type(fnames_to_label))
print(len(fnames_to_label))

to_label = table2_all_metadata.loc[table2_all_metadata['Filename'].isin(fnames_to_label)]
to_label.head()

to_label.nunique()

dict_df = pd.DataFrame(list(FileToLabel.items()), columns=['Filename', 'Class Label'])
dict_df.head()

OUTPUT = to_label.merge(dict_df)
OUTPUT.head()

print(OUTPUT.nunique())

print(OUTPUT['Class Label'].value_counts())

OUTPUT.to_csv('/Users/dh_lab_05/Desktop/FINAL_Labelling/Annotated_ECCO_I_57k.csv', index=None)

check = pd.read_csv('./Annotated_ECCO_I_57k.csv')
check.head()

check.nunique()
